# src/server.ts
import { prisma } from "../lib/prisma.js";
import express from "express";
import swaggerUi from "swagger-ui-express";
import swaggerDocument from "../swagger.json" with { type: "json" };

const port = 3000;
const app = express();

app.use(express.json());
app.use("/docs", swaggerUi.serve, swaggerUi.setup(swaggerDocument));

app.get("/movies", async (_, res) => {
    const movies = await prisma.movie.findMany({
        orderBy: {
            title: "asc",
        },
        include: {
            genres: true,
            languages: true,
        },
    });
    res.json(movies);
});

app.get("/movies/:id", async (req, res) => {
    const id = Number(req.params.id);

    try {
        const movies = await prisma.movie.findUnique({
            where: { id },
            include: {
                genres: true,
                languages: true
            }
        });

        if (!movies) {
            return res.status(404).send({ message: "Filme não encontrado" });
        }

        res.status(200).json(movies);
    } catch (error) {
        return res.status(500).send({ message: "Erro ao buscar o filme" });
    }
});

app.post("/movies", async (req, res) => {
    const { title, genre_id, language_id, oscar_count, release_date } =
        req.body;

    try {
        const movieWithSameTitle = await prisma.movie.findFirst({
            where: {
                title: { equals: title, mode: "insensitive" },
            },
        });

        if (movieWithSameTitle) {
            res.status(409).send({
                message: "Já existe um filme com este titulo",
            });
        }

        await prisma.movie.create({
            data: {
                title,
                genre_id,
                language_id,
                oscar_count,
                release_date: new Date(release_date),
            },
        });

        res.status(201).send();
    } catch (error) {
        return res.status(500).send({ message: "Erro ao cadastrar o filme" });
    }
});

app.put("/movies/:id", async (req, res) => {
    const id = Number(req.params.id);
    try {
        const movie = await prisma.movie.findUnique({
            where: { id },
        });

        if (!movie) {
            return res.status(404).send({ message: "Filme não encontrado" });
        }

        const data = { ...req.body };
        data.release_date = data.release_date
            ? new Date(data.release_date)
            : undefined;

        await prisma.movie.update({
            where: { id },
            data,
        });
    } catch (error) {
        return res
            .status(500)
            .send({ message: "Falha ao atualizar o registro do filme" });
    }

    res.status(200).send();
});

app.delete("/movies/:id", async (req, res) => {
    const id = Number(req.params.id);

    try {
        const movieExists = prisma.movie.findUnique({ where: { id } });

        if (!movieExists) {
            return res.status(404).send({ message: "Filme não encontrado" });
        }

        await prisma.movie.delete({ where: { id } });
    } catch (error) {
        res.status(500).send({ message: "Falha ao remover um filme" });
    }

    res.status(200).send();
});

app.get("/movies/:genreName", async (req, res) => {
    try {
        const moviesFilteredGenreName = await prisma.movie.findMany({
            include: {
                genres: true,
                languages: true,
            },
            where: {
                genres: {
                    name: {
                        equals: req.params.genreName,
                        mode: "insensitive",
                    },
                },
            },
        });
        res.status(200).send(moviesFilteredGenreName);
    } catch (error) {
        res.status(500).send({ message: "Falha ao encontrar o filme" });
    }
});

app.listen(port, () => {
    console.log(`Servidor em execução na porta: ${3000}`);
});